6 aliens shooting lasers
I strongly believe that there just isn't enough memory or that the Spartan 6 fpga doesn't have the capability to make the full game. I could be wrong, and perhaps my code style sucks, but I believe my implementation should work, but just isn't. 

As of now, I am able to get the spaceship to shoot lasers and destroy the flying saucer and the alien ships. In addition, I have gotten the alien ships to shoot lasers (based on different timers) and destroy the player spaceship